% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_re_genes.R
\name{annotate_with_coaccessibility}
\alias{annotate_with_coaccessibility}
\title{Annotate regulatory elements with gene targets using co-accessibility links}
\usage{
annotate_with_coaccessibility(links, re, txdb, name_links = "HPAP")
}
\arguments{
\item{links}{GInteractions with CCAN metadata}

\item{re}{GRanges of regulatory elements with 'region' column}

\item{txdb}{TxDb object for promoter extraction}

\item{name_links}{Character suffix for gene link metadata column (default "HPAP")}
}
\value{
GRanges of regulatory elements with un-nested gene annotations
}
\description{
Annotate regulatory elements with gene targets using co-accessibility links
}
