% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_spicey.R
\name{annotate_with_nearest}
\alias{annotate_with_nearest}
\title{Annotate RETSI peaks with nearest gene GETSI scores}
\usage{
annotate_with_nearest(retsi, getsi)
}
\arguments{
\item{retsi}{GRanges with RETSI scores and cell_type}

\item{getsi}{GRanges with GETSI scores, symbol, cell_type, and optionally entropy}
}
\value{
GRanges with RETSI and corresponding GETSI scores from nearest genes
}
\description{
Annotate RETSI peaks with nearest gene GETSI scores
}
