% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biomart_genes}
\alias{biomart_genes}
\title{Retrieve gene annotations from Ensembl via biomaRt}
\usage{
biomart_genes()
}
\value{
A list with two elements:
\describe{
\item{df}{A \code{data.frame} with gene annotation information for RNA-seq analysis.}
\item{gr}{A \code{GRanges} object with protein-coding genes for epigenomic analyses.}
}
}
\description{
Downloads gene annotation data from Ensembl using biomaRt, processes it,
and returns a list with both a data.frame and a filtered GRanges object
containing protein-coding genes. This is useful for RNA-seq and epigenomic
pipelines.
}
