% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsi.R
\name{compute_entropy_getsi}
\alias{compute_entropy_getsi}
\title{Compute normalized entropy of GETSI scores}
\usage{
compute_entropy_getsi(gr)
}
\arguments{
\item{gr}{GRanges with GETSI scores and symbol column}
}
\value{
Data frame with symbol and norm_entropy columns
}
\description{
Compute normalized entropy of GETSI scores
}
