% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.parse_input_diff}
\alias{.parse_input_diff}
\title{Parses input data of various types (e.g., named lists of \code{GRanges}
or \code{data.frame}, or a \code{GRangesList}) into a single tidy
\code{data.frame}, with a \code{cell_type} column.}
\usage{
.parse_input_diff(input)
}
\arguments{
\item{input}{An object representing differential results, such as:
\itemize{
\item A named list of \code{GRanges} objects.
\item A named list of \code{data.frame}s.
\item A \code{GRangesList}.
}}
}
\value{
A \code{data.frame} combining all elements, with an added \code{cell_type} column indicating the source.
}
\description{
Parses input data of various types (e.g., named lists of \code{GRanges}
or \code{data.frame}, or a \code{GRangesList}) into a single tidy
\code{data.frame}, with a \code{cell_type} column.
}
