% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spicey_indices.R
\name{entropy_index}
\alias{entropy_index}
\title{Compute normalized entropy of specificity scores}
\usage{
entropy_index(df, group_col)
}
\arguments{
\item{df}{Data.frame with columns for group_col, cell_type, and score_col.}

\item{group_col}{Character string name of grouping column ("gene_id" or "region").}
}
\value{
Data.frame with group_col and norm_entropy columns.
}
\description{
General function to compute entropy given a dataframe, grouping column, and score column.
}
