% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_re_genes.R
\name{get_ccan}
\alias{get_ccan}
\title{Extract CCAN membership from links for a given GRanges object}
\usage{
get_ccan(links, gr, name_column = "region", split = c("ccan", "name"))
}
\arguments{
\item{links}{GInteractions with CCAN1 and CCAN2 metadata columns}

\item{gr}{GRanges to query}

\item{name_column}{Character, metadata column in gr to use as identifier (default "region")}

\item{split}{Character, "ccan" returns names per CCAN, "name" returns CCANs per name}
}
\value{
Named list split by CCAN or name as requested
}
\description{
Extract CCAN membership from links for a given GRanges object
}
