% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_main_seqlevels}
\alias{get_main_seqlevels}
\title{Get main chromosomes from a GRanges or TxDb, filtering alt/hap/random/etc.}
\usage{
get_main_seqlevels(
  txdb,
  keep_mito = FALSE,
  include_only = NULL,
  verbose = FALSE
)
}
\arguments{
\item{keep_mito}{Logical, whether to keep mitochondrial/plastid contigs.}

\item{include_only}{Optional character vector of seqlevels to force-include (manual override).}

\item{verbose}{Logical, print what is kept/removed.}

\item{x}{A TxDb, GRanges, or similar object.}
}
\value{
Character vector of filtered seqlevels.
}
\description{
Get main chromosomes from a GRanges or TxDb, filtering alt/hap/random/etc.
}
