% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{granges_to_string}
\alias{granges_to_string}
\title{Converts a \code{GRanges} object into a character vector of strings, where each element
represents a genomic range in the format \code{"chr-start-end"}.}
\usage{
granges_to_string(gr)
}
\arguments{
\item{gr}{A \code{GRanges} object.}
}
\value{
A character vector, each element representing a range in \code{"chr-start-end"} format.
}
\description{
Converts a \code{GRanges} object into a character vector of strings, where each element
represents a genomic range in the format \code{"chr-start-end"}.
}
