% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_spicey.R
\name{link_spicey}
\alias{link_spicey}
\title{Link RETSI peaks with GETSI values based on gene linkage method}
\usage{
link_spicey(retsi_annotated, getsi, method = NULL)
}
\arguments{
\item{retsi_annotated}{A \code{data.frame} object containing RETSI results
including one of: \code{gene_coacc} (for coaccessibility) or \code{nearestGeneSymbol} (for nearest).}

\item{getsi}{A \code{data.frame} or tibble containing GETSI results with columns:
\itemize{
\item \code{gene_id}
\item \code{cell_type}
\item \code{GETSI}
\item \code{GETSI_entropy}
}}

\item{method}{Character; either \code{"coaccessibility"} or \code{"nearest"} to indicate
the gene linkage method used to associate RETSI regions with GETSI scores.
If \code{"coaccessibility"}, links are based on co-accessible genes (\code{gene_coacc});
if \code{"nearest"}, links are based on the nearest gene symbol (\code{nearestGeneSymbol}).}
}
\value{
A \code{data.frame} combining RETSI and GETSI specificity scores,
linked by gene identifier and including entropy values for each.
}
\description{
This generalized function links RETSI (Regulatory Element Tissue Specificity Index) regions
with GETSI (Gene Expression Tissue Specificity Index) scores based on either coaccessibility
or nearest gene annotation.
}
\examples{
data(atac)
data(rna)
data(cicero_links)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
getsi <- compute_spicey_index(rna=rna, gene_id = "gene_id")
retsi <- compute_spicey_index(atac=atac)

retsi_gene_nearest <- annotate_with_nearest(retsi = retsi,
                                            txdb = TxDb.Hsapiens.UCSC.hg38.knownGene,
                                            annot_dbi = org.Hs.eg.db::org.Hs.eg.db)
spicey_nearest <- link_spicey(retsi_annotated = retsi_gene_nearest,
                              getsi = getsi,
                              method = "nearest")

retsi_gene_coacc <- annotate_with_coaccessibility(links = cicero_links,
                                                  retsi = retsi,
                                                  txdb = TxDb.Hsapiens.UCSC.hg38.knownGene,
                                                  annot_dbi = org.Hs.eg.db::org.Hs.eg.db,
                                                  coaccess_cutoff_override = 0.25)
spicey_coacc <- link_spicey(
  retsi_annotated = retsi_gene_coacc,
  getsi = getsi,
  method = "coaccessibility"
)
}
