% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_spicey.R
\name{link_spicey_coaccessible}
\alias{link_spicey_coaccessible}
\title{Link RETSI peaks with GETSI values from coaccessibility}
\usage{
link_spicey_coaccessible(retsi_annotated_coacc, getsi)
}
\arguments{
\item{getsi}{A data.frame or tibble containing GETSI results with at least the columns:
\itemize{
\item \code{gene_id}: gene gene_id
\item \code{cell_type}
\item \code{GETSI}
\item \code{norm_entropy}
\item \code{seqnames}, \code{start}, \code{end}
}}

\item{re}{A data.frame or tibble containing RETSI results with at least the columns:
\itemize{
\item \code{gene_coacc}: gene symbols of the coaccessible gene to that region
\item \code{cell_type}
\item \code{norm_entropy}
\item \code{seqnames}, \code{start}, \code{end}, \code{annotation}, \code{distanceToTSS}, \code{region}, \code{RETSI}
}}
}
\value{
A data.frame with combined RETSI and GETSI information including entropy values,
gene annotations, and harmonized genomic coordinates.
}
\description{
This function combines GETSI (Gene Expression Tissue Specificity Index)
data with RETSI (Regulatory Element Tissue Specificity Index) results,
matching on gene symbols and cell types. It renames entropy columns,
performs joins to merge relevant annotations, and resolves coordinate conflicts.
}
