% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_spicey.R
\name{link_spicey_nearest}
\alias{link_spicey_nearest}
\title{Link RETSI peaks with GETSI values from nearest genes}
\usage{
link_spicey_nearest(retsi_annotated_nearest, getsi)
}
\arguments{
\item{retsi_annotated_nearest}{GRanges with RETSI scores and nearest genes.}

\item{getsi}{GRanges with GETSI scores by gene and cell type.}
}
\value{
Annotated GRanges with RETSI and matched GETSI scores.
}
\description{
Link RETSI peaks with GETSI values from nearest genes
}
