% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retsi.R
\name{retsi}
\alias{retsi}
\title{Compute RETSI scores on ATAC peaks}
\usage{
retsi(atac_da)
}
\arguments{
\item{atac_da}{GRanges object with differential accessibility info (avg_log2FC, p_val, cell_type columns)}
}
\value{
GRanges with computed RETSI scores (RETSI column)
}
\description{
Compute RETSI scores on ATAC peaks
}
