% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_re_genes.R
\name{annotate_with_nearest}
\alias{annotate_with_nearest}
\title{Annotates regulatory elements (e.g., ATAC-seq peaks) to the nearest gene
based on distance to the transcription start site (TSS), using a \code{TxDb}
reference and optional gene annotations from \code{org.*.db} packages.}
\usage{
annotate_with_nearest(
  peaks,
  txdb,
  annot_dbi,
  protein_coding_only = TRUE,
  verbose = TRUE,
  add_tss_annotation = FALSE,
  upstream,
  downstream
)
}
\arguments{
\item{peaks}{A \code{GRanges} or \code{data.frame} of peaks with at least the following columns:
\describe{
\item{seqnames}{Chromosome name of the regulatory region (e.g., \code{"chr1"}).}
\item{start}{Start coordinate of the peak.}
\item{end}{End coordinate of the peak.}
\item{region_id}{Unique identifier of the region (e.g., \code{chr1-5000-5800}).}
}}

\item{txdb}{\code{TxDb} object for genome annotation (required if annotation requested).}

\item{annot_dbi}{\code{AnnotationDbi} object for gene ID mapping (required if annotation requested).}

\item{protein_coding_only}{Logical; restrict to protein-coding genes (default TRUE).}

\item{add_tss_annotation}{Logical; annotate regulatory elements overlapping TSS (default FALSE).
If TRUE, use +/- 1bp TSS instead of +/-2kb promoter.}

\item{upstream}{Single integer value indicating the number of bases upstream
from the TSS (transcription start sites) (default 2000kb).}

\item{downstream}{Single integer values indicating the number of bases downstream
from the TSS (transcription start sites) (default 2000kb).}
}
\value{
A \code{data.frame} of peaks annotated to its nearest gene, with columns:
\itemize{
\item \code{distanceToTSS}: distance to the nearest TSS
\item \code{nearestGeneSymbol}: Official gene symbol of the nearest gene (e.g., GAPDH)
\item \code{annotation}: \code{"Promoter"} or \code{"Distal"} based on distance
}
}
\description{
Annotates regulatory elements (e.g., ATAC-seq peaks) to the nearest gene
based on distance to the transcription start site (TSS), using a \code{TxDb}
reference and optional gene annotations from \code{org.*.db} packages.
}
\examples{
library(dplyr)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
library(org.Hs.eg.db)

data(atac)
data(rna)
data(cicero_links)
peaks <- .parse_input_diff(atac)
peaks <- peaks \%>\% tidyr::separate(region_id,
                into = c("chr", "start", "end"), sep = "-",
                convert = TRUE,remove = FALSE) \%>\%
                GenomicRanges::makeGRangesFromDataFrame(keep.extra.columns = TRUE)

annotation_near <- annotate_with_nearest(peaks = peaks,
                                         txdb = TxDb.Hsapiens.UCSC.hg38.knownGene,
                                         annot_dbi = org.Hs.eg.db,
                                         protein_coding_only = TRUE,
                                         verbose = TRUE,
                                         add_tss_annotation = FALSE,
                                         upstream = 2000,
                                         downstream = 2000)
}
