% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cicero_links}
\alias{cicero_links}
\title{Example Cicero co-accessibility links}
\format{
A data frame with the following columns:
\describe{
\item{Peak1}{Genomic coordinate or identifier of the first peak}
\item{Peak2}{Genomic coordinate or identifier of the second peak}
\item{coaccess}{Co-accessibility score or correlation value between the two peaks}
}
Peaks must correspond exactly to those used in the ATAC-seq input.
}
\source{
Simulated Cicero results on example ATAC-seq data
}
\usage{
data(cicero_links)
}
\description{
A toy or example dataset representing co-accessibility links inferred from tools such as Cicero.
These links are used in SPICEY for associating regulatory elements to target genes.
}
\keyword{datasets}
