% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPICEY.R
\name{.parse_input_diff}
\alias{.parse_input_diff}
\title{Parse inputs from the different accepted types into data.frames}
\usage{
.parse_input_diff(input)
}
\description{
Parse inputs from the different accepted types into data.frames
}
