% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsi.R
\name{entropy_getsi}
\alias{entropy_getsi}
\title{Compute normalized entropy of GETSI scores}
\usage{
entropy_getsi(getsi_df)
}
\arguments{
\item{getsi_df}{Dataframe with GETSI scores and gene_id column}
}
\value{
Data frame with gene_id gene_id and norm_entropy columns
}
\description{
Compute normalized entropy of GETSI scores
}
