% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retsi.R
\name{entropy_retsi}
\alias{entropy_retsi}
\title{Compute normalized entropy of RETSI scores}
\usage{
entropy_retsi(retsi)
}
\arguments{
\item{gr}{GRanges with RETSI scores and region column}
}
\value{
Data frame with region and norm_entropy columns
}
\description{
Compute normalized entropy of RETSI scores
}
