% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_re_genes.R
\name{get_targets_links}
\alias{get_targets_links}
\title{Link regulatory elements to genes based on shared CCAN membership}
\usage{
get_targets_links(links, re, proms, name_links = "HPAP")
}
\arguments{
\item{links}{GInteractions with CCAN metadata}

\item{re}{GRanges of regulatory elements with 'region' column}

\item{proms}{GRanges of promoters with 'symbol' column}

\item{name_links}{Character suffix for metadata column name (default "HPAP")}
}
\value{
GRanges of regulatory elements annotated with linked genes
}
\description{
Link regulatory elements to genes based on shared CCAN membership
}
