% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPICEY.R
\name{link_spicey}
\alias{link_spicey}
\title{Link RETSI and GETSI}
\usage{
link_spicey(
  retsi = NULL,
  region_id = NULL,
  getsi = NULL,
  gene_id = NULL,
  annotation = NULL
)
}
\arguments{
\item{retsi}{RETSI calculated from DAR.}

\item{region_id}{A character string specifying the column name in each list element
that contains the accessible region identifiers.}

\item{getsi}{GETSI calculated from DEG.}

\item{gene_id}{A character string specifying the column name in each list element
that contains the gene identifiers.}

\item{annotation}{A data.frame linking \code{gene_id} to \code{region_id}. The should have
the same names provided in the respective parameters. This can be provided
by the user or generated using the function \code{\link{link_atac_to_genes}}.}
}
\value{
A \code{data.frame} combining RETSI and GETSI information
}
\description{
Link RETSI and GETSI
}
